requirements.txt
discord.py == 1.3.3
python_dotenv == 0.13.0